WIP Add vite support for SSR
