+++ examples/website/index.html
@@ -21,7 +21,7 @@
     </style>
   </head>
   <body>
-    <div id="root"></div>
-    <script type="module" src="src/index.ts"></script>
+    <!--ssr-outlet-->
+    <script type="module" src="/src/index.ts"></script>
   </body>
 </html>
