+++ examples/website/package.json
@@ -4,12 +4,16 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "serve": "vite preview"
+    "serve": "vite preview",
+    "ssr": "node src/server"
   },
   "devDependencies": {
     "autoprefixer": "^10.3.1",
     "postcss": "^8.3.6",
     "tailwindcss": "^2.2.7",
     "vite": "^2.4.3"
+  },
+  "dependencies": {
+    "express": "^4.17.1"
   }
 }
