+++ examples/website/src/components/SessionContext.ts
@@ -1,4 +1,4 @@
-import { createContext } from "../../../lib";
+import { createContext } from "../../../../lib";
 
 type Session = {
   token: string | null;
