+++ examples/website/src/entry-server.js
@@ -1,0 +1,8 @@
+import { renderToString, createElement } from "../../../lib";
+import App from "./components/App";
+
+export const render = (url) => {
+  const html = renderToString(createElement(App));
+  console.log({ html });
+  return html;
+};
